fboqrrequirements.txt
streamlit
qrcode
Pillow
pandas
